feat: LectureInfoStore 선언
- TimeTableStore에 너무 많은 정보가 담기는 문제
- 기능에 맞게 분리
