+++ client/src/stores/TimeTableStore.ts
@@ -13,7 +13,6 @@ interface TimeTableStoreState {
   tableIndex: ReactiveVar<number>;
   selectedTabLectures: ReactiveVar<Array<LectureInfos[]>>;
   colorIndex: ReactiveVar<number>;
-  lectures: ReactiveVar<LectureInfos[]>;
 }
 
 const colors = [
@@ -48,7 +47,6 @@ class TimeTableStore {
       tableIndex: makeVar(0),
       selectedTabLectures: makeVar<Array<LectureInfos[]>>([]),
       colorIndex: makeVar<number>(Math.floor(Math.random() * colors.length)),
-      lectures: makeVar<LectureInfos[]>(testData),
     };
   }
 
@@ -135,201 +133,5 @@ class TimeTableStore {
     return colors[colorIndex()];
   }
 }
-
-const testData = [
-  {
-    code: '111111',
-    name: '리눅스의 기초',
-    class: '01',
-    prof: '도눅스',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [
-      { start: 3690, end: 3720 },
-      { start: 540, end: 600 },
-    ],
-  },
-  {
-    code: '111112',
-    name: '신과함께',
-    class: '01',
-    prof: '갓우진',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 2340, end: 2460 }],
-  },
-  {
-    code: '222222',
-    name: '백엔드 심화과정',
-    class: '01',
-    prof: '백마',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 5100, end: 5160 }],
-  },
-  {
-    code: '333333',
-    name: '허접한 진혀쿠',
-    class: '01',
-    prof: '지녀쿠',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 5400, end: 5520 }],
-  },
-  {
-    code: '444444',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: '555555',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: '666666',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: '777777',
-    name: '디자인',
-    class: '01',
-    prof: '정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYs',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYa',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYb',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYc',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYd',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYe',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYf',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYg',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYh',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYi',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-  {
-    code: 'HANPYO',
-    name: '디자인커뮤니케이션',
-    class: '01',
-    prof: '윤정식',
-    grade: '03',
-    personnel: '25',
-    dept: '디자인건축공학부',
-    time: [{ start: 3660, end: 3780 }],
-  },
-];
 
 export default TimeTableStore;
