+++ client/src/stores/index.ts
@@ -2,4 +2,5 @@ export { default as RootStore } from './RootStore';
 export { default as ModalStore } from './ModalStore';
 export { default as TimeTableStore } from './TimeTableStore';
 export { default as SnackbarStore } from './SnackbarStore';
+export { default as LectureInfoStore } from './LectureInfoStore';
 export * from './StoreContext';
